feat(designtime): example of how designtime augments runtime
This is just a placeholder to show how the APIs should work and be constructed. In the end @aurelia/runtime and @aurelia/designtime will be two separate packages. The designtime package will depend on the runtime package. The build tooling will depend on the designtime package. When building an app, you will only need to ship the runtime. However, if you want progressive enhancement or need to do very dyanmic, on the fly view compilation, you would also ship with the designtime package.