feat(visual): making visual caching an implementation detail
TemplateController authors no longer need to worry about managing whether or not a visual is properly cached. Unbind will always return the visual to the cache and bind will ensure that it isn't in the cache. This all assumes that the developer is using caching. As part of this process, the interfaces related to Visual/IVisual and DefaultViewFactory/IViewFactory were refined.