deat(view-slot): shift responsibilities around
Remove all responsibilities for controlling binding from view slot. That must now be controlled by the owner of the slot that creates the views. This is necessary since different template controllers want to handle binding differently. An example is if, which binds to the current scope, while repeat binds to each item individually.

Removed all responsibilities for managing the attach/detach of views and viewSlots from teamplate controllers. That is now handled inside the attribute controller implementation and by the view slot.