refactor(all): first step of improving DI use throughout
Systematically removing all singletons so that the Aurelia instance is the composition root for everything related to a app. The goal is that you should be able to have multiple, completely isoated Aurelia apps on the same page and even if they use the same custom elements, it's possible to isolate template compilcation and all services between the two instances.