feat(requirejs): implemented plugins for require.js
This feature adds two require.js plugins: one for loading views and one for loading html-only components. The plugins work at build-time and at run-time to convert HTML into the data structures required by the jit. The structures take care of resolving imports for views but don't fully compile the templates. A hook in the runtime will allow the jit to plugin and jit the templates into the final structures. The rendering engine has been updated to detect this scenario. It's possible that these plugins could be enhanced in the future to do a full compile, either at build-time or run-time. This proves the basic idea behind extracting the loaders out of the framework and should work with system.js, Webpack and Parcel equally well (along with any loader/bundler technology that supports plugins).