doc(runtime): clarify the purpose of several core items
Also, added some notes on which things are likely to go away with the removal of shadow dom emulation. In the end, I'm hopeful that a good chunk of this will be simplified. However, I think it's likely going to still be very valuable to have core abstractions like ITemplates that create IViews. However, the simplicity can come from the fact that you will no longer need a RenderSlot to render the IView. You can basically just have an INode (or RenderLocation, which will just be an INode) and simply append the IView at the location.