feat(templating): generalize attribute and element attachable child
Previously, custom elements and custom attributes had a $slot property that could point to a render slot that was generated for template controller attributes or when something like compose asked for a slot representing itself. This change generalizes the mechanism by changing $slot to $child and switching to the more general IAttach interface, allowing custom attributes to connect an arbitrary attachable child into the attach lifecycle. In upcoming commits, this will enable a simplified implementation of replaceable and other single-child template controllers. Note that because IRenderSlot is an IAttach, advanced template controllers that manage multiple views still attach in the same as before.