feat(templating): convert replaceable to use render location (#59)
* feat(templating): generalize attribute and element attachable child

Previously, custom elements and custom attributes had a $slot property that could point to a render slot that was generated for template controller attributes or when something like compose asked for a slot representing itself. This change generalizes the mechanism by changing $slot to $child and switching to the more general IAttach interface, allowing custom attributes to connect an arbitrary attachable child into the attach lifecycle. In upcoming commits, this will enable a simplified implementation of replaceable and other single-child template controllers. Note that because IRenderSlot is an IAttach, advanced template controllers that manage multiple views still attach in the same as before.

* feat(templating): switch replaceable to use render location

* refactor(runtime): move test fakes to a shareable location

This is in preparation for writing some more tests for other templating resources.

* test(templating): add tests for replaceable attribute

* test(replaceable): change chai assertion method