refactor(styles): rename to make processor clear
There is a CSS Modules spec for native import of CSS files through the ES 2015 module system. There is also a CSS Modules preprocessor, which is not related to that spec. This commit renames things to make clear the support for the preprocessor. The actual CSS Modules spec is automatically supported via the Shadow DOM styles support.