fix(test/realworld): improve routerhook and route callbacks
The intention was to be able to start the realworld app. A couple of errors prevented this,
at first a type error in the authorisation router hook, and after that the runtime error
to check if the user is logged in. The callbacks are defined in the app class,
but the scope of the callback is in the navigation classes. The change fixes the scope for the callback.

Fix for #917