Removed confusing infobox and corrected incorrect lifecycle name
I removed the Aurelia Architecture info box under Component Registration. I could not find a reference to the `StandardConfiguration` export mentioned there. It feels like that section is referencing an older part of the docs.

Also, did a best-effort attempt to update the lifecycle hook name based on the description.