fix(new-instance): correctly invoke a registered interface
when using @newInstanceOf and @newInstanceForScope, even with an interface that already has a registered resolver, it will still throw as as it always tries to construct the interface, instead of the  corresponding implementation.

partially fixes #1166: this commit prepares a test where the most intuitive behavior is shown: ability to invoke an interface without having to declare it, if it has a default registration. Though this is inconsistent with the core, so will have to reconsider