refactor(switch+promise): deferred view instantiation
This fix addresses the reported issue on discourse
(https://discourse.aurelia.io/t/switch-case-statements-in-templates/4702)
of the CE VM are being instantiated even if the cases are not matched.
Previously the controller was created in the `Case#ctor`. This commit
defers it to the `Case#activate`; that is when the case matches the
switch condition and about to be activated.