chore(router-lite): removed Navigation
It is removed as it is purely for a marking purpose to indicate whether or not there is an active navigation. This might have served more purpose when there was direct routing involved, but now it is a simple marker. Thus, it removed and the boolean `_isNavigating` property is used instead.