refactor(router-lite): better path generation and parameter handling (#1495)
* docs(router-lite): engineering docs WIP

* docs(router-lite): engineering docs WIP

* feat(router-lite): most-matching path generation

* test: router*

* refactor(router-lite): path generation

* refactor(router-lite): route generation for non-string component

* docs(router-lite): engineering

* refactor(router-lite): routing-trigger

* refactor(router-lite): viewportinstruction cleanup

Removed the routing context from viewport instruction, as it was not
used. Additionally, removed the `IViewportInstruction`, as the
`ViewportInstruction` is by nature an intermediate object to derive the
RouteNode, and hence it makes little sense to inject it anywhere.
Consequently, the `ViewportInstruction`is also marked `@internal`.

* refactor(router-lite): path generation

BREAKING CHANGE
The support for `router.load(routeId, {params})` is removed in favor of `router.load({component: routeId, params})`. This is more inline with the existing infra of loading sibling routes: `router.load(['path/one', 'path/two'])`. The parallel in terms of using routeId and parameters would be `router.load([{component: 'r1', params: params1}, {component: 'r2', params: params2}])`.

* refactor(router-lite): support for CE,DEDef,VM,RtCfg in load

* docs(router-lite): wip

* fix(router-lite): eager route recognition

* test: router-lite

* test(router-lite): adding failing test

* chore(router-lite): cleanup

* fix(router-lite): broken test

* fix(router-lite): replace query for non-root ctx

Query params were not replaced when the navigation was triggered from a non-root context

* fix(router-lite): eager resolution for children

* test(router-lite): path generation for optional param

* chore(router-lite): cleanup

* docs(routing-lite): engineering

* chore(router-lite): removed Navigation

It is removed as it is purely for a marking purpose to indicate whether or not there is an active navigation. This might have served more purpose when there was direct routing involved, but now it is a simple marker. Thus, it removed and the boolean `_isNavigating` property is used instead.

* chore(router-lite): cleanup

* chore(router-lite): cleanup

* docs(routing-lite): engineering docs

* chore(router-lite): clean up append flag

Affected classes:
- RouteNode
- ViewportRequest
- ViewportInstruction
- NavigationOptions
- CompositeSegmentExpression

* fix(router-lite): tests

* chore(router-lite): minor changes

* chore(router-lite): minor changes - vpa

* chore(router-lite): minor changes - vpa

* docs(router-lite): jsdoc for component-agent

* refactor(router-lite): support for complex type in configured route data

* test(router-lite): lifecycle-hooks

* test(router-lite): lifecyclehooks

* test(router-lite): adding failing tests for lifecyclehooks

* chore(router-lite): temporarily skipped tests

* fix(router-lite): deepscan issues

* docs(router-lite): post review changes