fix(router-lite): handling hash in load, href CAs
The configuration options supported was supported to be a function
that starts the Router. This was undocumented, also probably contrived.
This is removed, and the options are initialized in the `hydrated` phase
rather than the `activated` phase which was the case previously.
This hindered the availability of the customization of options to the
CAs at the right time.

Note that this is a BREAKING CHANGE, although it might be relatively
low-impacting,as the "feature" was undocumented in the first place.