fix(router-lite): fallback with ce-name
Fixed the case of fallback when an empty route is defined. Previously it was so that when recognizing the ce-name, without any configured route with the same path, the empty route was selected. It is however expected that the correct CE definition is selected nevertheless.

This commit fixes the issue by comparing the fallback with the residue. If both matched, then it is assumed that fallback is matched with the empty route and no reasonable route-recognition took place. In such case, we can fall back to resolving the route definition by ce-name.