fix(router-lite/viewport): null default binding
Because the default value for the `default` bindable in the au-viewport is '', it allows loading the empty route by default. However, this results in a bit weird behavior, when there are sibling viewports with default value for the `default` property. In this case both the viewports load the empty route (assuming the empty route is configured). This commit allows binding null to the `default` property which indicates that the viewport must be ignored when free and no component should be loaded by default.

This commit additionally removes the resolution mode. This option changes only the lifecycle hook timings. IMO that is a very niche property, someone should be bothered about in devland. On the other hand, keeping this configuration property alive, hindered in the objective behind ignoring viewports for default components.