fix(router-lite): routeTree adjustment - canUnload
This PR fixes an edge case (maybe not so edge case).

Previously, it was such that when a `canUnload` returned `false` for an attempt to load `some/path`, another repeat attempt to load the same `some/path` was no-operational. This was due to the fact that even before the `canUnload` is invoked, the route tree was updated. Therefore, on the next attempt, the children node for the current tree and the nodes from the (retry) instruction matched, which instructs the router not to process the instruction.

This commit fixes that by marking unsuccessful unload and thereby enabling processing of the same routing instruction.