fix(router-lite): local dependencies registration
This commit fixes the duplicate registration of dependencies for routed view models. As the Controller.$el already does the same, the duplicate registrations are avoided from the router-lite. This was critical for the lifecycle hooks, as it caused 2 instances of every lifecycle hooks registered to the custom element's dependencies.

This change caused another side effect while resolving the custom element definition for fallbacks, using the custom element name. As the local dependencies are no longer registered to the container, resolving the fallback using the custom element name was no longer working as expected as it is possible that the fallback is not registered directly to this container. Thus as a solution, the `IContainer#find` now also takes a `searchAncestor` arg which can be set to true to traverse the container chain up.