fix(proxy-observation): prevent proxies from being wrapped in proxies again (#1716)
If you pass in a property that's part of a computed observation
structure into another class, the value is already wrapped in a proxy
before that property is observed in the new class. If that new class
then also uses computed observation, that property will be wrapped in a
proxy again, causing a "proxy within a proxy". This can repeat itself
endlessly, causing all sorts of issues.

Simply adding the proxy itself as the key to the `proxyMap` will solve
that problem at least as far as the issue in our app is concerned.