feat(router-lite): error recovery
This commit adds support for error recovery.  So far, if a routed component throws error while activating, for example error thrown from loading etc., the router-lite makes no attempt to recover from this error. It is the same if the navigation fails due to incorrect instructions (errors thrown internally from the router) etc. Thus, it delegates the responsibility to the devs. Though this strict behavior is not necessarily a bad thing, it is for most often undesirable.

This commit restores the previous route tree as the recovery step. Note that it triggers the lifecycle hooks and potentially makes expensive operations on dev land. Hence this behavior is kept under a router configuration flag `restorePreviousRouteTreeOnError`, which is set to `true` by default, and thus enabling error recovery. Set this to `false`, if a stricter behavior is desired. However, in that case, you need to ensure the avoidance of errors.