refactor(router-lite): miscellaneous changes (#1773)
* chore(router-lite): cleanup

* refactor(router-lite): residue handling

* fix(router-lite): viewport name match for contains check in RouteNode

* chore(router-lite): minor cleanup

* test(router-lite): instr with parenthesized param

* chore(router-lite): resolving todo in test

* test(router-lite): self-referencing config

* feat(router-lite): error recovery

This commit adds support for error recovery.  So far, if a routed component throws error while activating, for example error thrown from loading etc., the router-lite makes no attempt to recover from this error. It is the same if the navigation fails due to incorrect instructions (errors thrown internally from the router) etc. Thus, it delegates the responsibility to the devs. Though this strict behavior is not necessarily a bad thing, it is for most often undesirable.

This commit restores the previous route tree as the recovery step. Note that it triggers the lifecycle hooks and potentially makes expensive operations on dev land. Hence this behavior is kept under a router configuration flag `restorePreviousRouteTreeOnError`, which is set to `true` by default, and thus enabling error recovery. Set this to `false`, if a stricter behavior is desired. However, in that case, you need to ensure the avoidance of errors.

* test(router-lite): error recovery

* test(router-lite): error recovery

* test(router-lite): recovery from activation err

* test(router-lite): hierarchical error recovery

* chore(router-lite/tests): cleanup

* test: router-lite

* fix(router-lite): error recovery from child's hook

* test: router-lite

* test: router-lite

* test: router-lite

* test: router-lite

* chore(router-lite): bundle size optimization WIP

* refactor(router-lite): component-agent event trc

* refactor(router-lite): loc-mngr event log

* refactor(router-lite): logging

* refactor(router-lite): router event log

* refactor(router-lite): logging

* refactor(router-lite): logging

* refactor(router-lite): logging

* refactor(router-lite): error handling

* fix(router-lite): tests

* chore(router-lite/href): optimize for bundle size

* chore(router-lite/load): optimize for bundle size

* chore(router-lite/viewport): optimize for bundle size

* chore(router-lite/ca): optimize for bundle size

* refactor(router-lite/vi): removed obtrusive props from public API

* refactor(router-lite): optimize for bundle size

* refactor(router-lite): optimize for bundle size

* refactor(router-lite): optimize for bundle size

* refactor(router-lite): optimize for bundle size

* refactor(router-lite): optimize for bundle size

* refactor(router-lite): optimize for bundle size

* refactor(router-lite): optimize for bundle size

* fix(router-lite): deepscan issue

* chore(router-lite): cleanup

* chore(router/lite/events): documentation