chore(router-lite): removed IRouteNode implementation in RouteNode
The inheritance was relatively superfluous. The `IRouteNode` is also not a significant part of the public API. This commit corrects that by renaming the said interface to `IRouteNodeInitializationOptions` as the interface is merely used for creating a `RouteNode`. However, the interface is exported from the package as `RouteNode#create` is a part of the public API.