docs(template-compiler): extensive docs for template compiler errors (#2149)
* docs(kernel): extensive docs writing and refactoring for kernel docs

Some of the existing errors were incorrect and there were a lot of missing errors.

Furthermore, the `createMappedError` method has been modified to link to the official Aurelia docs when the error is thrown in the console (allowing easier access to errors).

* docs(dialog): add dedicated dialog error pages

Adding in error pages for the dialog plugin.

Furthermore, the `createMappedError` method has been modified to link to the official Aurelia docs when the error is thrown in the console (allowing easier access to errors).

* fix(test): dialog service test updated to reflect new error message

Now we have added in a link to the error code in the docs, the previous test failing as it was throwing more than just an error message.

* docs(template-compiler): extensive docs for template compiler errors

Documenting template compiler error message codes into dedicated error doc pages.

Furthermore, the `createMappedError` method has been modified to link to the official Aurelia docs when the error is thrown in the console (allowing easier access to errors).

* refactor(packages): only link to docs for errors in dev mode

* refactor(template-compiler): only link to errors in dev mode