feat(runtime-html): add value converter context (#2159)
* feat(runtime-html): add value converter context

Introduces a mechanism for value converters to access contextual information, such as the target element, component, and binding instance.

This enables more sophisticated value converter logic that can adapt based on the environment in which it is used, allowing for more flexible and reusable converters, but primarily motivated by the desire to be able to debug where value converters are being run.

* refactor(binding): value converter context resolution

Improves value converter context resolution by using a lazy context creation mechanism.
This ensures that context properties like target and binding are always available.

Removed other fluff properties from the context as well as unnecessary checks and assertions.

* refactor(binding): value converter caller context

Updates value converters to receive caller context as a second argument when `useCallerContext` is set to true.

Removes the `ICallerContextResolver` service and simplifies the way caller context is accessed, improving performance and reducing complexity. The binding is now passed as the second argument, containing `target`, `bridge` and `source` information.

* refactor(binding): check controller is an instance of controller

* refactor(binding): caller context retrieval

Simplifies caller context retrieval in bindings by using the hydration context.

This change improves the reliability and efficiency of accessing the component context for binding expressions. It ensures the correct component is always used as the binding source, regardless of the DOM structure. It removes a heap of boilerplate code too.

* refactor(binding): contextualise value converter caller

Updates value converters to use a `contextual` flag instead of `useCallerContext` for determining whether to pass the binding context.

This change provides a more explicit and consistent way to manage the caller context within value converters.

* refactor(binding): rename contextual vc caller properties

Renames contextual value converter caller properties for clarity.

Instead of `target` and `bridge`, the caller now exposes `source` and `binding`, respectively, providing more relevant context to the value converter. Target is considered unstandardised, so removing it is safer.

The tests are updated to reflect this change, ensuring the correct context is passed in different scenarios.

* refactor(binding): refine value converter caller context

Standardizes the type for the caller context passed to contextual value converters.

This change improves type safety and consistency by introducing a specific `ICallerContext` interface, which is now used consistently across the value converter implementation.

* refactor(binding): add additional tests for args and context

Tests to ensure that instances where params and context is enabled work correctly, also properly asserting fromView.

* refactor(binding): simplify value converter execution

Avoids unnecessary try-catch block when obtaining hydration context. Adds early exits to value converter execution when toView/fromView are not defined.

* refactor(binding): value converter context handling

Updates value converter context handling to use `withContext` property instead of `contextual` property, simplifying the logic and improving readability.

This change ensures that value converters correctly receive the caller context when needed, providing access to the binding and source information.

Tests and use of fixture methods also have been used to cleanup boilerplate.

* feat(binding): adds au-slot context tests for value converters

Adds tests to verify correct caller context (source and binding) is provided to value converters used within `au-slot`, including scenarios with named slots and repeaters.

* refactor(binding): remove stop call from fixture usage in context tests

As per PR feedback, we don't need to call stop as things are handled for us in the fixture creation.

* refactor(binding): positive checks first

As per PR feedback, favour positive checks over negative ones.

* refactor(binding): remove withcontext from decorator definition

This is unused now

* docs(value-converters): add docs for using the new caller context

* fix: update chromedriver

---------

Co-authored-by: bigopon <bigopon.777@gmail.com>