feat(runtime-html): allow components in interpolation
This allows rendering components in views using interpolation as discussed in issue #2111 - I did quickly look at passing the component using the React style syntax, until the caveats listed in the linked issue became immediately obvious if you wanted bindable properties.

This was a new part of the codebase for me, so it might not be 100% concrete. But it works and follows similar rendering location logic that we use in `au-compose`