docs(errors): lots and lots and lots of error pages for linking
this has been on the boiler for a long time now and it has taken forever to get to all these. I've been working on them on and off, but realised a lot were still missing. I did these over the course of a week, it was a lot. Also made sure packages like router were linking to said errors.